requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
fontawesomefree==6.2.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
